requirements.txt
python-ucto
regex
lxml
tqdm
pandas
folia
glob
langdetect